Fixing errors

+++ Raycaster/Raytracer/VoxelScene.cpp
@@ -739,9 +739,9 @@ void VoxelScene::simplify()
 	}
 }
 
-void VoxelScene::simplifyOctree(Octree<glm::dvec3>* tree)
+void VoxelScene::simplifyOctree(Octree<glm::vec3>* tree)
 {
-	glm::dvec3* color = nullptr;
+	glm::vec3* color = nullptr;
 
 	//Check if leaves are similar
 	for (uint8_t i = 0; i < 8; i++)
@@ -776,7 +776,7 @@ void VoxelScene::simplifyOctree(Octree<glm::dvec3>* tree)
 	delete[] tree->tree;
 
 	tree->contains = OCTREE_CONTENT::FILLED;
-	tree->object = new glm::dvec3(*color);
+	tree->object = new glm::vec3(*color);
 
 	//std::cout << "merged\n";
 }
